Refactor : 메인 컴포넌트 index로 네이밍 변경
